src/__tests__/nips/nip39.test.ts
import { describe, it, expect } from 'vitest';
import { mockSecp256k1, mockBech32, mockBip39, TEST_CONSTANTS } from '../test-utils';
import {
  generateSeedWords,
  validateSeedWords,
  encryptSeedPhrase,
  decryptSeedPhrase
} from '../../nips/nip39';

// Setup mocks
mockSecp256k1();
mockBech32();
mockBip39();

describe('NIP-39: Client Behaviors', () => {
  describe('Seed Phrase Generation', () => {
    it('should generate valid seed phrases', () => {
      const seedPhrase = generateSeedWords({ name: 'Test User' });
      expect(seedPhrase).toBeDefined();
      expect(typeof seedPhrase).toBe('string');
      expect(seedPhrase.split(' ').length).toBe(12);
    });

    it('should validate correct seed phrases', () => {
      const seedPhrase = generateSeedWords({ name: 'Test User' });
      const isValid = validateSeedWords(seedPhrase.split(' '));
      expect(isValid).toBe(true);
    });

    it('should reject invalid seed phrases', () => {
      const invalidPhrases = [
        '',
        'invalid',
        'too short phrase',
        'this is not a valid seed phrase at all',
        '12345 67890 not valid'
      ];

      invalidPhrases.forEach(phrase => {
        const isValid = validateSeedWords(phrase.split(' '));
        expect(isValid).toBe(false);
      });
    });
  });

  describe('Seed Phrase Encryption', () => {
    const testPassword = 'test-password-123';

    it('should encrypt seed phrases', async () => {
      const seedPhrase = generateSeedWords({ name: 'Test User' });
      const encrypted = await encryptSeedPhrase(seedPhrase, testPassword);
      
      expect(encrypted).toBeDefined();
      expect(typeof encrypted).toBe('string');
      expect(encrypted).not.toBe(seedPhrase);
    });

    it('should decrypt seed phrases correctly', async () => {
      const originalPhrase = generateSeedWords({ name: 'Test User' });
      const encrypted = await encryptSeedPhrase(originalPhrase, testPassword);
      const decrypted = await decryptSeedPhrase(encrypted, testPassword);
      
      expect(decrypted).toBe(originalPhrase);
    });

    it('should fail decryption with wrong password', async () => {
      const seedPhrase = generateSeedWords({ name: 'Test User' });
      const encrypted = await encryptSeedPhrase(seedPhrase, testPassword);
      
      await expect(decryptSeedPhrase(encrypted, 'wrong-password'))
        .rejects.toThrow();
    });
  });

  describe('Error Handling', () => {
    it('should handle empty passwords', async () => {
      const seedPhrase = generateSeedWords({ name: 'Test User' });
      await expect(encryptSeedPhrase(seedPhrase, ''))
        .rejects.toThrow();
    });

    it('should handle empty seed phrases', async () => {
      await expect(encryptSeedPhrase('', 'password'))
        .rejects.toThrow();
    });

    it('should handle invalid encrypted data format', async () => {
      await expect(decryptSeedPhrase('invalid-data-string', 'password'))
        .rejects.toThrow();
      await expect(decryptSeedPhrase('invalid-data', 'password'))
        .rejects.toThrow();
    });
  });

  describe('Security Requirements', () => {
    it('should use secure encryption', async () => {
      const seedPhrase = generateSeedWords({ name: 'Test User' });
      const encrypted = await encryptSeedPhrase(seedPhrase, 'password');
      
      // Encrypted data should be significantly longer than input
      expect((await encrypted).length).toBeGreaterThan(seedPhrase.length);
      
      // Encrypted data should not contain the original phrase
      expect((await encrypted).includes(seedPhrase)).toBe(false);
    });

    it('should generate different ciphertexts for same input', async () => {
      const seedPhrase = generateSeedWords({ name: 'Test User' });
      const password = 'test-password';
      
      const encrypted1 = await encryptSeedPhrase(seedPhrase, password);
      const encrypted2 = await encryptSeedPhrase(seedPhrase, password);
      
      // Should use random IV/salt, so ciphertexts should be different
      expect(encrypted1).not.toBe(encrypted2);
    });
  });
});
